Add try catch for reading incoming packets
It seems there was quite often a packet coming in that could not be
read, it gave RangeError’s when trying to read it. By catching it we at
least prevent the code from crashing when such a packet comes in.
Further examination is needed to figure out what is “wrong” with these
incoming packets.